Move git info to tail of previous command
Editing in emacs may change the status of the git repo without requiring
a command to be executed in the shell.  So, the git status info really
applies to the state of the repo just after executing the previous
command, not to the state of the repo as the current command is
executed.  This location reflects that better.